Modeled the data and IO of an XCC file.
Binding the data to the XML data is done using the Jackson XML library. Created various serializer and deserializer for handling custom data, such as MeasurementUnitDeserializer.kt, MeasuringPresetSerializer.kt, and MeasuringPresetDeserializer.kt. Added an accumulator to integrate the measured data. Added a .gitignore file to prevent specific files and directories from being included in the repository.