Updated pom.xml, added Kotlin dependencies and peeteew module
The major changes in this commit include an update to the groupId in pom.xml, addition of Kotlin-related dependencies and configurations, compiler target and source versions have been set to 8, and the addition of a new module named 'peeteew'. Additionally, the corresponding pom.xml file is created for 'peeteew' module.